Set "-bash [name]" as sandbox-use proctitle.
Where "bash" is actually the basename of their shell, falling back to sh and "name" is the name of the sandbox being used, surrounded by brackets just like in /etc/profile.d/sandbox_prompt.sh.

This is the first step on the long and possibly impossible road to shrinking the process tree.